Finished AudioSettingPanel, added AddUserPanel, added new Audio Formats
User is now also able to choose an output device, can control the output
volume and select the type he wishes to use to save his recorded voice
messages as.
Also added a panel that lets you type in the username of a user you want
to add to your contact list and displays all the available contacts the
server sends, which fit to the entered username.