Implement profile system. Closes #35
With this commit the profile system was finally implemented. The server
can receive SendableProfiles from the client and update the client's
profile.
Furthermore a SendableRequest was introduced which can be used to
request a profile update from the server. It also replaces the now
deprecated SendableDeleteAccount. Detailed usage information through
JavaDoc has to be added via a future commit.
All enums were updated with proper formatting.
DAOs and Managers were updated to work with the new ProfileSystem.
New Exceptions for Profiles have been added.
A few methods (especially in Server) have been commented.