Manage Client Settings on Server. Closes #36
Further development on the UuidEntity and Manger system introduced in
the previous commit. This includes subtypes of UuidEntity and
EntityManger to further generalize a lot of code. Profiles and Settings
are now processed mostly in the same superclass.
This commit also introduces new Sendables: for Settings and for
UuidEntities in general (internal reasons).
Updated Settings class.
Updated Hibernate configuration.
Updated unit tests.
Updated and improved many other things, too.