New Message System - WIP! Updates #82. Includes Fix for #10
With the new message system it will be possible to store Messages
permanently. Messages are now a subtype of UuidEntity, and can therefore
be processed similarly to profiles and settings. The SendableMessage now
holds a Message object instead of the raw messsage data. Specific
Messages don't need their own Sendable anymore.
The new system makes chat histories possible and is a lot tidier
overall.

Warning: Not all unit tests are working right now. There seems to be a severe problem with referencing the messages in the database within a sendable.